git-svn-id: svn://192.168.1.10/public/CoffeeMud@12772 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Areas/StdPlanet.java
@@ -61,8 +61,8 @@ public class StdPlanet extends StdTimeZone implements SpaceObject
 	}
 	@Override public double[] direction(){return emptyDirection;}
 	@Override public void setDirection(double[] dir){ }
-	@Override public long speed(){return 0;}
-	@Override public void setSpeed(long v){}
+	@Override public double speed(){return 0;}
+	@Override public void setSpeed(double v){}
 	@Override public long radius() { return radius; }
 	@Override public void setRadius(long radius) { this.radius=radius; }
 	@Override public void setName(String newName)
